+++ src/vue.d.ts
@@ -7,6 +7,7 @@ import contextMenu from '@/components/ContextMenu'
 import { resetForm } from '@/utils/form'
 import { notImplemented } from '@/utils/notImplemented'
 import { copy } from '@/utils/clipboard'
+import { sort } from '@/utils/sortChange'
 
 // 在 Vue 实例上扩展自定义属性 https://vuejs.org/guide/typescript/options-api.html#augmenting-global-properties
 declare module 'vue' {
@@ -31,6 +32,7 @@ declare module 'vue' {
     $resetForm: typeof resetForm
     $notImplemented: typeof notImplemented
     $copy: typeof copy
+    $sort: typeof sort
   }
 }