+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 
  * Add sync database table column
  * Add monaco editor
+ * Support global sort query